refactor: 로또 컨트롤러로 분리
